Build Vignettes for Github

This function aims at processing vignette (building, organizing) before pushing on Github so that user can download a package with pre-compiled vignette. Using the website argument allow to build the associated website.

Usage

build_github_vignette(
  article = NULL,
  website = NULL,
  vignette = NULL,
  clean = FALSE
)

Arguments

article

The name or the path to the .Rmd file to be compiled

website

A boolean to compile the website associated with the article

vignette

A boolean to compile the vignette associated with the article

clean

Logical to remove cache and temporary files

Value

A \docs folder containing the website and a inst\doc folder containing the vignette

Details

To make sure pre-compiled vignette are enclosed within the package downloaded from Github, vignette are (i) compiled with tools::buildVignettes, (ii) then copy-paste to inst\doc. To build the website, this function first converts README.Rmd file to README.md file, and then calls pkgdown::build_site() since this function only take into account *.md.

References

https://community.rstudio.com/t/how-to-add-article-html-or-r-files-to-a-github-rep/45905/7

Examples

if (FALSE) {
# compile all articles into vignette and website
build_github_vignette()
}